Renaming package to option
