+++ vector/src/main/java/im/vector/app/features/location/option/LocationSharingOptionPickerView.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.location.view
+package im.vector.app.features.location.option
 
 import android.content.Context
 import android.util.AttributeSet
@@ -23,7 +23,6 @@ import android.view.View
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.view.isVisible
 import im.vector.app.databinding.ViewLocationSharingOptionPickerBinding
-import im.vector.app.features.location.LocationSharingOption
 
 /**
  * Custom view to display the location sharing option picker.
